Move unit tests on fim_db_files to its own file
